requirements.txt
crawl4ai>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.1
python-dotenv>=1.0.0 